Move payment type into signer constants
